Fixed apply() on ProbabilityDistribution

+++ src/main/scala/nlp/a1/ProbabilityDistribution.scala
@@ -5,9 +5,8 @@ import nlp.a0.Process
 
 class ProbabilityDistribution[B](labels: Map[B,Int],features: Map[B,Map[B,Map[B,Int]]]) extends ProbabilityDistributionToImplement[B] {
 	def apply(x: B): Double = {
-		val total = labels.foldLeft(0)(_+_._2) 
-		val label = labels(x)
-		// val out = (label/total) - ((label/total)%0.01)
+		val total = labels.foldLeft(0)(_+_._2).toDouble
+		val label = labels(x).toDouble
 		return label/total
 	}
